+++ src/modules/formBuilder/components/itemEditor/types.ts
@@ -1,0 +1,4 @@
+import { DeepPartial } from 'react-hook-form';
+import { FormItem } from '@/types/gqlTypes';
+
+export type FormItemState = DeepPartial<FormItem>;
